Create CQ Quiz App with 2 category of quiz
Nowadays, the multiple-choice question (MCQ) is very popular because of its ease of use. Evaluation is very easy in this system and the exam can be taken from anyplace and anytime. Your task is to create a MCQ Quiz app which will show the user a set of questions and the user will be able to participate in the examination.

Requirement and specification:

1. Show the question set and prompt user to type an answer. 2. After answering all the questions user will be shown total marks he obtained.

3. Exit the program when user types 0 and play again if he types 1